rutas recetas comprobadas

+++ routes/ingredient.routes.js
@@ -2,16 +2,18 @@ const express = require("express")
 const router = express.Router()
 
 const Ingredients = require("../models/Ingredients.model")
+const verifyToken = require("../middlewares/auth.middleware")
+const Recipe = require("../models/Recipe.model")
 
 //Crear un nuevo ingrediente SOLO ADMIN
-router.post(("/"), async(req, res, next) => {
+router.post(("/"),verifyToken, async(req, res, next) => {
 
     try {
         const response = await Ingredients.create({
             nombre: req.body.nombre,
             establecimiento: req.body.establecimiento,
             hidratos: req.body.hidratos,
-            creador: req.body. creador
+            creador: req.payload._id
         })
         res.json(response)
     } catch (error) {
@@ -22,14 +24,14 @@ router.post(("/"), async(req, res, next) => {
 
 //Editar un ingrediente SOLO ADMIN
 
-router.put("/:ingredientsId", async (req, res, next) => {
+router.put("/:ingredientsId",verifyToken, async (req, res, next) => {
 
     try {
         const responseFromDB = await Ingredients.findByIdAndUpdate(req.params.ingredientsId, {
             nombre: req.body.nombre,
             establecimiento: req.body.establecimiento,
             hidratos: req.body.hidratos,
-            creador: req.body. creador
+            creador: req.payload._id
         })
         res.json(responseFromDB)
     } catch (error) {
@@ -39,10 +41,17 @@ router.put("/:ingredientsId", async (req, res, next) => {
 })
 
 //Eliminar un ingrediente SOLO ADMIN
-//-verificar si existe ingr. en alguna receta, si existe no se puede eliminar
+//verificar si existe ingr. en alguna receta, si existe no se puede eliminar
 
-router.delete("/:ingredientsId", async(req, res, next) => {
+router.delete("/:ingredientsId",verifyToken, async(req, res, next) => {
     try {
+
+        const {ingredientsId} = req.params;
+        const ingExist = await Recipe.exists({Ingredients: ingredientsId});
+        if (ingExist) {
+            return res.send("No se puede eliminar este ingrediente, está siendo usado en alguna receta.")
+        }
+
         await Ingredients.findByIdAndDelete(req.params.ingredientsId)
         res.send("Ingrediente eliminado correctamente")
     } catch (error) {
